Time: 0 ms (100.00%) | Memory: 13.3 MB (96.19%) - LeetSync

+++ 1716-maximum-non-negative-product-in-a-matrix/maximum-non-negative-product-in-a-matrix.cpp
@@ -3,7 +3,7 @@ class Solution {
 public:
     static int maxProductPath(vector<vector<int>>& grid) {
         const int r=grid.size(), c=grid[0].size(), MOD=1e9+7;
-        int64_t p=dp[0][0][0]=dp[0][0][1]=grid[0][0];
+        long long p=dp[0][0][0]=dp[0][0][1]=grid[0][0];
         for(int j=1; j<c; j++){
             p*=grid[0][j];
             dp[0][j][0]=dp[0][j][1]=p;
@@ -18,7 +18,7 @@ public:
                 dp[i][j][0]=minP, dp[i][j][1]=maxP;
             }
         }
-        int64_t ans=dp[r-1][c-1][1];
+        long long ans=dp[r-1][c-1][1];
         return ans<0?-1: ans%MOD;
     }
 };
